+++ sys/vm/device_pager.pas
@@ -25,7 +25,7 @@ uses
  vm_pmap,
  sys_conf;
 
-function OFF_TO_IDX(x:QWORD):DWORD; inline;
+function OFF_TO_IDX(x:QWORD):QWORD; inline;
 begin
  Result:=QWORD(x) shr PAGE_SHIFT;
 end;
